+++ .gitignore
@@ -1,2 +1,2 @@
 /venv/
-
+/yolov5/
