+++ HealthManager.py
@@ -1,7 +1,6 @@
 from types import NoneType
 
-import speaker as spk
-
+from Modules import speaker as spk
 
 
 class HealthManager:
