+++ main.py
@@ -1,5 +1,4 @@
 import time
-from multiprocessing import Process
 
 import detectors
