--- other/get_audio_devices.py
@@ -1,3 +1,0 @@
-import sounddevice
-
-print(sounddevice.query_devices())
